requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
crewai[google-genai]
composio
composio-crewai
python-dotenv
